+++ src/main/java/cart/service/ProductService.java
@@ -24,12 +24,22 @@ public class ProductService {
     }
 
     public void updateProduct(final long id, final String name, final int price, final String imageUrl) {
+        validateExistData(id);
+
         final Product newProduct = new Product(id, name, price, imageUrl);
 
         productDao.update(newProduct);
     }
 
     public void deleteProduct(final long id) {
+        validateExistData(id);
+        
         productDao.delete(id);
+    }
+
+    private void validateExistData(final Long id) {
+        if (!productDao.isExist(id)) {
+            throw new IllegalArgumentException("존재하지 않는 id입니다.");
+        }
     }
 }
