+++ src/test/java/cart/dao/ProductDaoTest.java
@@ -1,7 +1,6 @@
 package cart.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import cart.domain.Product;
@@ -62,13 +61,6 @@ class ProductDaoTest {
     }
 
     @Test
-    void 존재하지_않는_상품_조회시_예외_발생() {
-        assertThatIllegalArgumentException().isThrownBy(
-                () -> productDao.findById(10L)
-        ).withMessage("존재하지 않는 id입니다.");
-    }
-
-    @Test
     void 상품_데이터_수정() {
         final Long id = 2L;
         final Product newProduct = new Product(id, "new salad", 3000, "url");
@@ -83,14 +75,18 @@ class ProductDaoTest {
     }
 
     @Test
+    void 존재하지_않는_상품_수정시_예외_발생() {
+        final Product notExistProduct = new Product(3L, "커피", 2800, "매머드 커피");
+
+        productDao.update(notExistProduct);
+    }
+
+    @Test
     void 상품_데이터_삭제() {
         final Long id = 2L;
 
         productDao.delete(id);
 
-        assertThatIllegalArgumentException().isThrownBy(
-                () -> productDao.findById(id)
-        ).withMessage("존재하지 않는 id입니다.");
+        assertThat(productDao.findAll().size()).isEqualTo(1);
     }
-
 }
